requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.0.0
Pillow>=10.0.0
